requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
ultralytics==8.3.100
ultralytics-thop==2.0.14
pillow==10.2.0
opencv-python==4.11.0.86
streamlit==1.35.0
requests==2.32.3
python-multipart==0.0.20
numpy==1.25.2
matplotlib==3.10.1
pandas==2.2.3
seaborn==0.13.2
scipy==1.15.2
torch==2.6.0
torchvision==0.21.0